requirements.txt
Package          Version
---------------- -----------
boto3            1.40.73
botocore         1.40.73
iniconfig        2.3.0
jmespath         1.0.1
libcachesim      0.3.3.post3
numpy            2.3.4
packaging        25.0
pip              25.3
pluggy           1.6.0
pybind11         3.0.1
pybind11-stubgen 2.5.5
pygments         2.19.2
pytest           9.0.1
python-dateutil  2.9.0.post0
pyyaml           6.0.3
s3transfer       0.14.0
six              1.17.0
tqdm             4.67.1
urllib3          2.5.0
